Criado procesador de contexto para passar user para todas las plantillas; eliminada ruta logar_usuario.
