+++ flask_app/controllers/autentication_controller.py
@@ -26,7 +26,7 @@ def login_usuario():
     session['nombre'] = usuario.nombre
     session['apellido'] = usuario.apellido
     session['categoria'] = usuario.categoria
-    return redirect("/logar_usuario")
+    return redirect("/")
 
 @app.route('/close_session')
 def close_session():
