+++ flask_app/templates/components/_menu_lateral.html
@@ -7,7 +7,7 @@
   <hr>
   <ul class="nav nav-pills flex-column mb-auto">
     <li class="nav-item">
-      <a href="/logar_usuario" class="nav-link active" aria-current="page">
+      <a href="/" class="nav-link active" aria-current="page">
         <svg class="bi me-2" width="16" height="16">
           <use xlink:href="#home"/>
         </svg>
@@ -62,10 +62,12 @@
     <a href="#" class="d-flex align-items-center text-white text-decoration-none dropdown-toggle" id="dropdownUser1"
       data-bs-toggle="dropdown" aria-expanded="false">
       <i class="fas fa-user-circle fa-2x me-2"></i>
-      <strong>{{usuario | default('Usuario') }}</strong>
+      {% if usuario != None %}
+      <strong>{{usuario}}</strong>
+      {% endif %}
     </a>
     <ul class="dropdown-menu dropdown-menu-dark text-small shadow" aria-labelledby="dropdownUser1">
-      {% if usuario is not defined %}
+      {% if usuario == None %}
       <li><a class="dropdown-item" href="#" data-bs-toggle="modal" data-bs-target="#registroModal"><i
             class="fas fa-user me-2"></i>Registro</a></li>
       <li>
@@ -77,7 +79,7 @@
         <hr class="dropdown-divider">
       </li>
       {% endif %}
-      {% if usuario is defined %}
+      {% if usuario != None %}
       <li><a class="dropdown-item" href="/close_session"><i class="fas fa-sign-out-alt me-2"></i>Cerrar Sesión</a>
       </li>
       {% endif %}
