+++ flask_app/utils/__init__.py
@@ -1,0 +1,6 @@
+## Coding Dojo - Python Bootcamp Ene 2025
+## Proyecto final
+## Wavely
+ 
+ 
+# Este archivo vacio es necesário para que sea posible utilizar el 'procesador de contexto'
